docker/requirements.txt
datasets==2.21.0
pandas==2.2.2
#protobuf==5.27.3
protobuf==4.25.4
pymilvus==2.4.5
streamlit==1.37.1
torch==2.4.0
transformers==4.44.1
google.generativeai==0.7.2
